api/pyproject.toml
[tool.poetry]
name = "backend-exercise"
version = "0.1.0"
description = ""
authors = ["Mike Liu <mike@bearwavetech.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
djangorestframework-api-key = "^2.1.0"
djangorestframework-simplejwt = "^5.0.0"
django-cryptography = "^1.0"
django-filter = "^21.1"
django-guardian = "^2.4.0"
django-extensions = "^3.2.3"
drf-yasg = "^1.20.0"
uvicorn = "^0.17.4"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
